make evio single event buffer lengths an unsigned 32 bit int instead of 64 bits

+++ src/libsrc++/evioBufferChannel.cc
@@ -160,7 +160,7 @@ bool evioBufferChannel::read(uint32_t *myBuf, int length) throw(evioException) {
  *
  * Note:  user MUST free the allocated buffer!
  */
-bool evioBufferChannel::readAlloc(uint32_t **buffer, uint64_t *bufLen) throw(evioException) {
+bool evioBufferChannel::readAlloc(uint32_t **buffer, uint32_t *bufLen) throw(evioException) {
   noCopyBuf=NULL;
   if(handle==0)throw(evioException(0,"evioBufferChannel::readAlloc...0 handle",__FILE__,__FUNCTION__,__LINE__));
 
@@ -187,7 +187,7 @@ bool evioBufferChannel::readAlloc(uint32_t **buffer, uint64_t *bufLen) throw(evi
 bool evioBufferChannel::readNoCopy(void) throw(evioException) {
   if(handle==0)throw(evioException(0,"evioBufferChannel::readNoCopy...0 handle",__FILE__,__FUNCTION__,__LINE__));
 
-  uint64_t bufLen;
+  uint32_t bufLen;
   int stat=evReadNoCopy(handle,&noCopyBuf,&bufLen);
   if(stat==EOF)return(false);
   if(stat!=S_SUCCESS) throw(evioException(stat,"evioBufferChannel::readNoCopy...read error: " + string(evPerror(stat)),
